MergeRequest.people_involved was returning a wrong list of users.
When the user commented on its own merge request the server was
breaking and presenting only a white page. This was due to a hard
coded value used instead of the id for the merge request, which
caused an empty list to be returned, and in the end the "to" field
for the email was set to "nil", over there on MergeRequestMailer.updated.