New algorithm to calculate path diff.
The output now is similar to what git outputs, at least for all
(not so many) cases I saw.

Start comparing the paths in reversal order and stop at a difference,
if two blocks of differences were found, better render two paths
instead of a lot of {foo => bar} blocks.