Split MR index in two sections.
First section is MRs that need your attention, the second are the MRs
that are waiting for others to do something.

MRs waiting for others are MRs that have the last comment in the last
patch made by myself or were created by myself and are without comments
in their last patch.